+++ swift-forabank/Sources/Landing/LandingUIComponent/Components/Carousel/Sizes.swift
@@ -9,10 +9,10 @@ import Foundation
 
 public struct Sizes: Equatable {
     
-    public let width: CGFloat
-    public let height: CGFloat
+    public let width: Int
+    public let height: Int
     
-    public init(width: CGFloat, height: CGFloat) {
+    public init(width: Int, height: Int) {
         self.width = width
         self.height = height
     }
